[MDC-666] simet-ma: twamp: use appropriately sized types
The internal timestamps in microsseconds MUST be a 64-bit type, since
they have the same epoch as NTP (i.e. they count the microsseconds since
the current NTP epoch).

Also, we must ensure the conversion routines use the larger 64-bit data
types internally, as well.

This fixes the TWAMP reports to match the in-wire absolute time.

+++ twamp-client-c/src/report.h
@@ -52,12 +52,12 @@ typedef struct report_packet {
     unsigned int reflectorSeqNumber;
     unsigned int receiverSeqNumber;
 
-    unsigned int senderTime_us;
-    unsigned int reflectorRecvTime_us;
-    unsigned int reflectorSendTime_us;
-    unsigned int receiverTime_us;
+    uint64_t senderTime_us;
+    uint64_t reflectorRecvTime_us;
+    uint64_t reflectorSendTime_us;
+    uint64_t receiverTime_us;
 
-    unsigned int rtt_us;
+    uint64_t rtt_us;
 } ReportPacket;
 
 #endif /* REPORT_H_ */
